refactor: 서비스 구조 수정으로 인한 import 변경
