+++ app/api/feedBack/route.ts
@@ -1,7 +1,8 @@
 import { NextResponse } from 'next/server';
 
 import { GenerateFeedbackAnswerRequest } from '@/app/_types/api/interview';
-import { generateFeedbackAnswer } from '@/services/interview';
+
+import { generateFeedbackAnswer } from '@/services/interview/service';
 
 // API 엔드포인트 로직
 export async function POST(request: Request) {
