+++ app/api/questions/another/route.ts
@@ -1,7 +1,8 @@
 import { NextResponse } from 'next/server';
 
 import { GenerateAnotherQuestionRequest } from '@/app/_types/api/interview';
-import { generateAnotherQuestion } from '@/services/interview';
+
+import { generateAnotherQuestion } from '@/services/interview/service';
 
 // API 엔드포인트 로직
 export async function POST(request: Request) {
